LayoutTests/platform/chromium-win/fast/loader/stateobjects/state-url-sets-links-visited-expected.txt

CONSOLE MESSAGE: line 14: Uncaught TypeError: Object [object Object] has no method 'computedStyleIncludingVisitedInfo'
This link should get colored visited as a result of replaceState() setting it as the current URL
This link should get colored visited as a result of pushState() adding it to the forward list
If you're running in a browser, the link should be orange-on-black and you should see "replacedURL.html" in your global history.
If you're running in DRT, the test will also append "PASS" or "FAIL".
